// components/service-pages/types.ts
import type { LucideIcon } from "lucide-react";

export interface ServiceOfferingItem {
  num: string;
  title: string;
  description: string;
  icon: LucideIcon;
}

export interface UseCaseItem {
  title: string;
  description: string;
  icon: LucideIcon;
}

export interface WhyIITILItem {
  num: string;
  title: string;
  description: string;
  icon: LucideIcon;
}

export interface BenefitItem {
  title: string;
  description: string;
  icon: LucideIcon;
}